Adding thresholding on GPU

+++ src/process.hpp
@@ -8,4 +8,4 @@ void average_pooling(const uint8_t* sobel_x, const uint8_t* sobel_y,
                      uint8_t *output, int width, int height,
                      int stride, int pool_size);
 
-void threshold(const uint8_t* buffer, uint8_t *output, int width, int height, int stride, int value);
+void threshold(const uint8_t* buffer, uint8_t *output, int width, int height, int stride, uint8_t value);
